+++ src/main/java/neobis/cms/Service/Bishkek/BishCoursesServiceImpl.java
@@ -1,9 +1,14 @@
 package neobis.cms.Service.Bishkek;
 
 import neobis.cms.Dto.CoursesDTO;
+import neobis.cms.Entity.Bishkek.BishClient;
 import neobis.cms.Entity.Bishkek.BishCourses;
+import neobis.cms.Entity.Bishkek.BishTeachers;
+import neobis.cms.Exception.IllegalArgumentException;
 import neobis.cms.Exception.ResourceNotFoundException;
+import neobis.cms.Repo.Bishkek.BishClientRepo;
 import neobis.cms.Repo.Bishkek.BishCoursesRepo;
+import neobis.cms.Repo.Bishkek.BishTeacherRepo;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,9 +16,13 @@ import java.util.List;
 @Service
 public class BishCoursesServiceImpl implements BishCoursesService{
     private final BishCoursesRepo coursesRepo;
+    private final BishTeacherRepo teacherRepo;
+    private final BishClientRepo clientRepo;
 
-    public BishCoursesServiceImpl(BishCoursesRepo coursesRepo) {
+    public BishCoursesServiceImpl(BishCoursesRepo coursesRepo, BishTeacherRepo teacherRepo, BishClientRepo clientRepo) {
         this.coursesRepo = coursesRepo;
+        this.teacherRepo = teacherRepo;
+        this.clientRepo = clientRepo;
     }
 
     @Override
@@ -61,6 +70,13 @@ public class BishCoursesServiceImpl implements BishCoursesService{
     public String deleteCourse(List<Long> courses) {
         for (long courseID : courses) {
             BishCourses course = this.findCourseById(courseID);
+            List<BishClient> bishClients = clientRepo.findAllByCourse(course);
+            if (!bishClients.isEmpty())
+                throw new IllegalArgumentException("Курс " + course.getName() + "используется, не может быть удален.");
+
+            BishTeachers teacher = course.getTeacher();
+            teacher.setCourse(null);
+            teacherRepo.save(teacher);
             coursesRepo.delete(course);
         }
         if (courses.size() > 1)
